Begin refactoring Database.jsx
