+++ client/src/routes.jsx
@@ -6,7 +6,7 @@ import Report from "./pages/Report.jsx";
 import Admin from "./Admin.jsx";
 import Login from "./pages-admin/Login.jsx";
 import Dashboard from "./pages-admin/Dashboard.jsx";
-import Database from "./pages-admin/Database.jsx";
+import Database from "./pages-admin/Database/Database.jsx";
 
 const routes = [
   {
